URL and MySQL Injection Protection

+++ admin/includes/view_all_users.php
@@ -54,10 +54,12 @@
     }
 
     if (isset($_GET['delete'])) {
-        $user_id = $_GET['delete'];
-        $query = "DELETE FROM users WHERE user_id = $user_id";
-        $delete_user_query = mysqli_query($connection, $query);
-        header("Location: users.php");
+        if(isset($_SESSION['user_role']) && $_SESSION['user_role'] == 'admin'){
+            $user_id = mysqli_real_escape_string($connection , $_GET['delete']);
+            $query = "DELETE FROM users WHERE user_id = $user_id";
+            $delete_user_query = mysqli_query($connection, $query);
+            header("Location: users.php");
+        }
     }
     ?>
     </tbody>
